@gitlawr 2020-08-17
By reducing the number of masters the testing duration drops to ~41s.
The following is the comparison diagram:
![compare-new](https://user-images.githubusercontent.com/5697937/90407960-bbb5cb80-e0d9-11ea-8df2-07ef7a02e900.png)

@liggitt  I've updated the pr per your suggestion