@liggitt 2020-08-17
/lgtm
/approve

thanks! I still think we should look for ways to improve this in the future, but this is a big initial improvement for a minimal change