@liggitt 2020-08-17
/milestone v1.19

this package timeout is currently the [primary source of flakes](https://storage.googleapis.com/k8s-gubernator/triage/index.html?ci=0&pr=1&job=integration&xtest=run_kubectl_apply_tests%7CTestLimits) in the pull-integration job